Move Base metrics related REST TCK tests to the optional module
