--- tck/rest/src/main/resources/application_metrics.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<!--
-    Copyright (c) 2017, 2022 Contributors to the Eclipse Foundation
-
-    See the NOTICES file(s) distributed with this work for additional
-    information regarding copyright ownership.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    You may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-    SPDX-License-Identifier: Apache-2.0
--->
-<config>
-  <metric multi="false" name="org.eclipse.microprofile.metrics.test.MetricAppBean.redCount" type="counter" unit="none" description="red-description"/>
-  <metric multi="false" name="org.eclipse.microprofile.metrics.test.MetricAppBean.blue" type="counter" unit="none"/>
-  <metric multi="false" name="greenCount" type="counter" unit="none"/>
-  <metric multi="false" name="purple" type="counter" unit="none" tags="app=myShop"/>
-  <metric multi="false" name="jellybeanHistogram" type="histogram" unit="jellybeans" description="jellybeans-description"/>
-  <metric multi="false" name="metricTest.test1.countMeA" type="counter" unit="none" description="count-me-a-description"/>
-  <metric multi="false" name="metricTest.test1.countMeB" type="counter" unit="jellybean"/>
-  <metric multi="false" name="org.eclipse.microprofile.metrics.test.MetricAppBean.gaugeMeA" type="gauge" unit="kibibits" description="gauge-me-a-description" />
-  <metric multi="false" name="org.eclipse.microprofile.metrics.test.MetricAppBean.gaugeMeB" type="gauge" unit="hands"/>
-  <metric multi="false" name="org.eclipse.microprofile.metrics.test.MetricAppBean.timeMeA" type="timer" unit="nanoseconds"/>
-</config>
