--- tck/rest/src/main/resources/base_metrics.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<!--
-    Copyright (c) 2017, 2022 Contributors to the Eclipse Foundation
-
-    See the NOTICES file(s) distributed with this work for additional
-    information regarding copyright ownership.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    You may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-    SPDX-License-Identifier: Apache-2.0
--->
-<config>
-  <metric multi="false" name="thread.count" type="gauge" unit="none"/>
-  <metric multi="false" name="thread.max.count" type="gauge" unit="none"/>
-  <metric multi="false" name="thread.daemon.count" type="gauge" unit="none"/>
-  <metric multi="false" name="memory.maxHeap" type="gauge" unit="bytes"/>
-  <metric multi="false" name="memory.usedHeap" type="gauge" unit="bytes"/>
-  <metric multi="false" name="memory.committedHeap" type="gauge" unit="bytes"/>
-  <metric multi="false" name="classloader.loadedClasses.count" type="gauge" unit="none"/>
-  <metric multi="false" name="classloader.loadedClasses.total" type="counter" unit="none"/>
-  <metric multi="false" name="classloader.unloadedClasses.total" type="counter" unit="none"/>
-  <metric multi="false" name="cpu.availableProcessors" type="gauge" unit="none"/>
-  <metric multi="false" name="jvm.uptime" type="gauge" unit="seconds"/>
-  <metric multi="true" name="gc.total" type="counter" unit="none" tags="name=*"/>
-  <metric multi="true" name="gc.time" type="gauge" unit="seconds" tags="name=*"/>
-  <metric multi="false" name="optional.metric" type="histogram" unit="seconds" optional="true"/>
-  <metric multi="false" name="cpu.systemLoadAverage" type="gauge" unit="none" optional="true"/>
-  <metric multi="false" name="cpu.processCpuLoad" type="gauge" unit="percent" optional="true"/>
-  <metric multi="false" name="cpu.processCpuTime" type="gauge" unit="seconds" optional="true"/>
-</config>
